Split Page into subcomponents
