+++ client/components/ScrollableContentWrapper.tsx
@@ -1,24 +1,32 @@
-import React, { CSSProperties, FC } from 'react';
+import React, { CSSProperties, forwardRef, memo, PropsWithChildren } from 'react';
 import SimpleBar from 'simplebar-react';
 import 'simplebar/src/simplebar.css';
 
 import { useDir } from '../hooks/useDir';
 
-
 const styleDefault = {
-	maxHeight: '100%', flexGrow: 1,
-};
+	maxHeight: '100%',
+	flexGrow: 1,
+} as const;
 
-
-type CustomScrollbarsProps = {
+type ScrollableContentWrapperProps = PropsWithChildren<{
 	onScroll?: Function;
-	ref: React.Ref<unknown>;
 	style?: CSSProperties;
-}
+}>;
 
-const ScrollableContentWrapper: FC<CustomScrollbarsProps> = React.memo(React.forwardRef(({ onScroll, children, style = styleDefault }, ref) => {
+const ScrollableContentWrapper = forwardRef<HTMLElement, ScrollableContentWrapperProps>(({
+	children,
+	style = styleDefault,
+	onScroll,
+}, ref) => {
 	const dir = useDir();
-	return <SimpleBar data-simplebar-direction={dir} direction={dir} style={style} scrollableNodeProps={{ ref, onScroll }} children={children}/>;
-}));
+	return <SimpleBar
+		data-simplebar-direction={dir}
+		direction={dir}
+		style={style}
+		scrollableNodeProps={{ ref, onScroll }}
+		children={children}
+	/>;
+});
 
-export default ScrollableContentWrapper;
+export default memo(ScrollableContentWrapper);
